Split out the tests
