+++ test/init.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const MochaApplication = require('../init/MochaApplication');
+const app = new MochaApplication();
+app.boot();
+
+module.exports = app;
